pyproject.toml
[tool.poetry]
name = "tets"
version = "0.1.0"
description = "build this hristmastreefarm Inventory Management - Provides tools to manage tree stock, track inventory levels, and update statuses, including items like fertilizer, dirt, saplings, hoses, trucks, harvesters, lights, etc. Sales Tracking - Track sales data, analyze trends, and integrate with QuickBooks for financial management. Scheduling - Manage planting, harvesting, and delivery schedules. Customer Management - Maintain customer records, preferences, and order history integrated with Quickbooks. Order Management - Streamline order processing, from placement to delivery, integrated with QuickBooks for invoicing. Supply Chain Management - Oversees the supply chain from seedling purchase to delivery of trees. Reporting and Analytics - Generate detailed reports and analytics to support business decisions, directly linked with QuickBooks for accurate financial reporting. Mapping and Field Management - Map farm layouts, manage field assignments and track conditions of specific areas. Health Management - Monitor the health of the trees and schedule treatments. Staff Roles Management - Define roles, responsibilities, and permissions for all staff members. Staff Scheduling - Manage schedules for staff operations, ensuring coverage and efficiency. Staff Performance Management - Evaluate staff performance, set objectives, and provide feedback. Payroll Management - Automate payroll calculations, adhere to tax policies, and integrate with QuickBooks. QuickBooks Integration - Integrate seamlessly across all financial aspects of the app to ensure comprehensive financial management."
authors = ["AutoGPT <info@agpt.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
bcrypt = "^3.2.0"
fastapi = "*"
passlib = {version = "^1.7.4", extras = ["bcrypt"]}
prisma = "*"
pydantic = "*"
pyjwt = "^2.5.0"
python-jose = "^3.3.0"
uvicorn = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
